adds a generic common button to common
