+++ src/components/common/Buttons/Button.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+import { Button } from 'antd';
+
+const StyledLink = styled(Link)`
+  text-align: center;
+`;
+
+const StyledButton = styled(Button)`
+  background-color: 293845;
+  width: 200px;
+  text-align: center;
+  margin: 20px auto;
+`;
+export const StyledBtn = ({ path, onClick }) => {
+  return (
+    <StyledButton size="large" type="primary" onClick={onClick}>
+      <StyledLink to={path}>Confirm</StyledLink>
+    </StyledButton>
+  );
+};
