+++ src/components/common/index.js
@@ -1,3 +1,4 @@
+import { StyledBtn } from './Buttons/Button';
 import FormButton from './FormButton';
 import FormInput from './FormInput';
 import List from './List';
@@ -8,6 +9,7 @@ import Button from './Button';
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
 export {
+  StyledBtn,
   FormButton,
   FormInput,
   LayoutContainer,
